* commit 'c6aa83125c22f64f32126e9ca6fe4de049d9bf3e':
  Cleaned up blt's CodeCoverage, per code review.
  Installation of docs and coverage no longer fails if directory to delete does not exist.
  Changes groups and permissions on web version of coverage results.
  Enabled code coverage in Fortran. ATK-688.
  Adds a verification check to ensure that CMake successfully generates the Fortran/C interface. ATK-741.
  Fixed inconsistency in build system's code coverage variable. ATK-739.

+++ scripts/bamboo/chaos-gcc@4.9.3-release.sh
@@ -49,16 +49,20 @@ cd ..
 
 echo "Installing docs to web space..."
 echo "-----------------------------------------------------------------------"
-rm -rf /usr/global/web-pages/lc/www/toolkit/docs_old
-if [ $? -ne 0 ]; then
-    echo "Error: 'rm' failed"
-    exit 1
+if [ -d /usr/global/web-pages/lc/www/toolkit/docs_old ]; then
+    rm -rf /usr/global/web-pages/lc/www/toolkit/docs_old
+    if [ $? -ne 0 ]; then
+        echo "Error: 'rm' of docs_old failed"
+        #exit 1
+    fi
 fi
 
-mv /usr/global/web-pages/lc/www/toolkit/docs /usr/global/web-pages/lc/www/toolkit/docs_old
-if [ $? -ne 0 ]; then
-    echo "Error: 'mv' failed"
-    exit 1
+if [ -d /usr/global/web-pages/lc/www/toolkit/docs ]; then
+    mv /usr/global/web-pages/lc/www/toolkit/docs /usr/global/web-pages/lc/www/toolkit/docs_old
+    if [ $? -ne 0 ]; then
+        echo "Error: 'mv' docs to docs_old failed"
+        # exit 1
+    fi
 fi
 
 cp -R ./install-chaos-release/docs /usr/global/web-pages/lc/www/toolkit/
